backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
chromadb==0.5.5
ollama==0.3.3
langchain-community==0.3.0
langchain==0.3.0
pymupdf==1.24.10
python-docx==1.1.2
watchdog==4.0.1
apscheduler==3.10.4
pydantic-settings==2.5.2
python-multipart==0.0.9
aiofiles==24.1.0
httpx==0.27.2
